code/UnMICST-P/requirements.txt
absl-py==0.9.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
boto3==1.14.32
botocore==1.17.32
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==1.4.1
cycler==0.10.0
cytoolz==0.10.1
dask==2.18.1
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
future==0.18.2
google-auth==1.16.1
google-auth-oauthlib==0.4.1
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imagecodecs==2020.5.30
imagecodecs-lite==2019.12.3
imageio==2.8.0
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
javabridge==1.0.19
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.5
jupyterlab-server==1.1.5
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.3.0.36
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydensecrf==1.0rc3
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-bioformats==4.0.0
python-dateutil==2.8.1
python-javabridge==4.0.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-image==0.17.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
terminado==0.8.3
testpath==0.4.4
tifffile==2020.6.3
toolz==0.10.0
torch==1.5.0+cu92
torchvision==0.6.0+cu92
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.1.0
